LayoutTests/imported/w3c/web-platform-tests/dom/nodes/aria-element-reflection.tentative-expected.txt

Item 1
Item 2
Item 1
Item 2
Hello world!

  Invalid Time Password:
First description.
Second description.
Item 1
Item 2
Item 1
Item 2
Billing
Name

Address

Tab 1
Tab 2
Fruit
Delicious
Nutritious
Title

Next
Content2
Content1
Parent
Child 1
Child 2
Light DOM Heading

Light DOM text

Light DOM text

Delicious
Nutritious

Misspelling
Wonderful
Fantastic


FAIL aria-activedescendant element reflection assert_equals: invalid ID for relationship returns null expected (object) null but got (undefined) undefined
FAIL If the content attribute is set directly, the IDL attribute getter always returns the first element whose ID matches the content attribute. assert_equals: expected (object) Element node <div role="option" id="option1">Item 1</div> but got (undefined) undefined
FAIL Setting the IDL attribute to an element which is not the first element in DOM order with its ID causes the content attribute to be an empty string assert_true: expected true got false
FAIL Setting an element reference that crosses into a shadow tree is disallowed, but setting one that is in a shadow inclusive ancestor is allowed. assert_equals: expected (object) null but got (undefined) undefined
FAIL aria-errormessage assert_equals: expected (string) "errorMessage" but got (object) null
FAIL aria-details assert_equals: expected (object) null but got (undefined) undefined
FAIL Deleting a reflected element should return null for the IDL attribute and cause the content attribute to become stale. assert_equals: expected (object) Element node <div role="option" id="contentAttrElement">Item 1</div> but got (undefined) undefined
FAIL Changing the ID of an element causes the content attribute to become out of sync. assert_equals: expected (object) Element node <div role="option" id="changingIdElement">Item 1</div> but got (undefined) undefined
FAIL Reparenting an element into a descendant shadow scope nullifies the element reference. assert_equals: expected (string) "lightElement" but got (object) null
FAIL aria-labelledby. assert_array_equals: parsed content attribute sets element references. value is undefined, expected array
FAIL aria-controls. assert_array_equals: value is undefined, expected array
FAIL aria-describedby. assert_array_equals: value is undefined, expected array
FAIL aria-flowto. assert_array_equals: value is undefined, expected array
FAIL aria-owns. assert_array_equals: value is undefined, expected array
FAIL shadow DOM behaviour for FrozenArray element reflection. assert_equals: expected (object) null but got (undefined) undefined
FAIL Moving explicitly set elements across shadow DOM boundaries. assert_equals: expected (string) "buttonDescription1 buttonDescription2" but got (object) null
FAIL Moving explicitly set elements around within the same scope, and removing from the DOM. assert_array_equals: aria-labeled by is supported by IDL getter. value is undefined, expected array
